Refactored the shim tests to extract test bins
