+++ shims/libpcisysfs/src/shim.c
